pyproject.toml
[tool.poetry]
name = "your-project-name"
version = "0.1.0"
description = "FastAPI server for web scraping and AI processing"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "^0.7.0"
anyio = "^4.6.2"
beautifulsoup4 = "^4.12.3"
bs4 = "^0.0.2"
certifi = "^2024.8.30"
charset-normalizer = "^3.4.0"
click = "^8.1.7"
distro = "^1.9.0"
dnspython = "^2.7.0"
duckduckgo-search = "^6.3.3"
email-validator = "^2.2.0"
fastapi = "^0.115.4"
fastapi-cli = "^0.0.5"
gitpython = "^3.1.43"
httpcore = "^1.0.6"
httptools = "^0.6.4"
httpx = "^0.27.2"
jinja2 = "^3.1.4"
jiter = "^0.7.0"
numpy = "^2.1.2"
ollama = "^0.3.3"
openai = "^1.53.0"
pgvector = "^0.3.6"
phidata = "^2.5.27"
primp = "^0.6.5"
psycopg = "^3.2.3"
psycopg2-binary = "^2.9.10"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
python-dotenv = "^1.0.1"
python-multipart = "^0.0.17"
pyyaml = "^6.0.2"
requests = "^2.32.3"
rich = "^13.9.3"
shellingham = "^1.5.4"
sqlalchemy = "^2.0.36"
starlette = "^0.41.2"
tinydb = "^4.8.2"
tomli = "^2.0.2"
tqdm = "^4.66.6"
typer = "^0.12.5"
typing-extensions = "^4.12.2"
urllib3 = "^2.2.3"
uvicorn = "^0.32.0"
uvloop = "^0.21.0"
watchfiles = "^0.24.0"
websockets = "^13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
